refactor: rm Input directory
