+++ packages/bezier-react/src/components/Forms/TextField/TextField.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/mixins/typography";
+@use "../../../styles/mixins/typography";
 
 .TextFieldInput {
   @include typography.size(14);
