+++ packages/bezier-react/src/index.ts
@@ -52,9 +52,9 @@ export * from '~/src/components/Forms/FormControl'
 export * from '~/src/components/Forms/FormGroup'
 export * from '~/src/components/Forms/FormHelperText'
 export * from '~/src/components/Forms/FormLabel'
-export * from '~/src/components/Forms/Inputs/Select'
-export * from '~/src/components/Forms/Inputs/TextArea'
-export * from '~/src/components/Forms/Inputs/TextField'
+export * from '~/src/components/Forms/Select'
+export * from '~/src/components/Forms/TextArea'
+export * from '~/src/components/Forms/TextField'
 export * from '~/src/components/Forms/RadioGroup'
 export * from '~/src/components/Forms/SegmentedControl'
 export * from '~/src/components/Forms/Slider'
